% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.daily.R
\name{load.daily}
\alias{load.daily}
\title{Get the daily load profiles.}
\usage{
load.daily(loadsample, load.date, num.obs)
}
\arguments{
\item{loadsample}{A numeric vector that demonstrates load profile of a household.
It can be a yearly/seasonally/monthly or even several days' load vector.}

\item{load.date}{A vector of characters with the form 'yyyy-mm-dd', which presents
the corresponding date vectors of the "loadsample".}

\item{num.obs}{An integer which presents the number of observations on each day.}
}
\value{
loadsample --- A numeric matrix with each row presents a daily observations.
}
\description{
Transfer multiple days of load vector of one household into one matrix with
each row presents a daily observation.
}
\examples{
### Import the load profiles.
#
data("loadsample")
#
### Get the load profile matrix of the household X1.
#
loadsample.X1<-load.daily(loadsample$X1, load.date=loadsample$date, num.obs=48)



}
\keyword{load}
